refactor(v2): organize state and instructions
