--- programs/omnipair-v2/src/instructions/hedge/mod.rs
@@ -1,5 +1,0 @@
-mod close_hedge;
-mod open_hedge;
-
-pub use close_hedge::*;
-pub use open_hedge::*;
