+++ programs/omnipair-v2/src/instructions/lending/mod.rs
@@ -1,10 +1,12 @@
 mod borrow;
 mod common;
 mod deposit_collateral;
+mod liquidation;
 mod repay;
 mod withdraw_collateral;
 
 pub use borrow::*;
 pub use deposit_collateral::*;
+pub use liquidation::*;
 pub use repay::*;
 pub use withdraw_collateral::*;
