+++ programs/omnipair-v2/src/instructions/liquidity/hlp/mod.rs
@@ -1,0 +1,5 @@
+mod deposit_single_sided;
+mod withdraw_single_sided;
+
+pub use deposit_single_sided::*;
+pub use withdraw_single_sided::*;
