+++ programs/omnipair-v2/src/instructions/liquidity/mod.rs
@@ -1,5 +1,9 @@
 mod claim_yield;
+mod hlp;
 mod set_yield_recipient;
+mod ylp;
 
 pub use claim_yield::*;
+pub use hlp::*;
 pub use set_yield_recipient::*;
+pub use ylp::*;
