+++ programs/omnipair-v2/src/instructions/liquidity/set_yield_recipient.rs
