+++ programs/omnipair-v2/src/instructions/mod.rs
@@ -1,19 +1,13 @@
 mod common;
 mod futarchy;
-mod hedge;
 mod lending;
-mod liquidation;
+mod liquidity;
 mod market;
-mod reserve;
 mod spot;
 pub mod transfer_hook;
-mod yielding;
 
 pub use futarchy::*;
-pub use hedge::*;
 pub use lending::*;
-pub use liquidation::*;
+pub use liquidity::*;
 pub use market::*;
-pub use reserve::*;
 pub use spot::*;
-pub use yielding::*;
