+++ programs/omnipair-v2/src/state/mod.rs
@@ -1,32 +1,11 @@
-pub mod config;
-pub mod debt;
-pub mod fees;
 pub mod futarchy_authority;
-pub mod health;
-pub mod hlp;
-pub mod insurance;
-pub mod limits;
 pub mod liquidation_auction;
+pub mod margin_position;
 pub mod market;
-pub mod positions;
-pub mod reserves;
-pub mod risk;
-pub mod shares;
-pub mod side;
 pub mod yield_account;
 
-pub use config::*;
-pub use debt::*;
-pub use fees::*;
 pub use futarchy_authority::*;
-pub use hlp::*;
-pub use insurance::*;
-pub use limits::*;
 pub use liquidation_auction::*;
+pub use margin_position::*;
 pub use market::*;
-pub use positions::*;
-pub use reserves::*;
-pub use risk::*;
-pub use shares::*;
-pub use side::*;
 pub use yield_account::*;
