--- programs/omnipair-v2/src/transitions/mod.rs
@@ -1,8 +1,0 @@
-pub mod collateral;
-pub mod debt;
-pub mod fee;
-pub mod hedge;
-pub mod interest;
-pub mod liquidation;
-pub mod reserve;
-pub mod swap;
